Refactor code
Added separated code for each component in main page -> footer,header,NewsCard,SentimentChart
