+++ src/components/Footer.js
@@ -1,0 +1,14 @@
+// src/components/Footer.js
+import React from "react";
+
+function Footer() {
+  return (
+    <footer className="bg-gradient-to-r from-blue-600 to-indigo-700 text-white py-8 text-center">
+      <p className="text-lg">
+        &copy; 2025 Crypto Sentiment. All rights reserved.
+      </p>
+    </footer>
+  );
+}
+
+export default Footer;
