+++ src/components/Header.js
@@ -1,0 +1,16 @@
+// src/components/Header.js
+import React from "react";
+
+function Header() {
+  return (
+    <header className="bg-gradient-to-r from-blue-600 to-indigo-700 text-white py-10 shadow-lg">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 text-center">
+        <h1 className="text-5xl font-extrabold tracking-tight text-transparent bg-clip-text bg-gradient-to-r from-teal-300 to-green-400">
+          Crypto News and Sentiment Analysis
+        </h1>
+      </div>
+    </header>
+  );
+}
+
+export default Header;
